Add files via upload
This is titled "pre-final" because I am still not sure if any errors will occur. I upload it without making it good-looking to read and apologize for any inconvenience it may cause. This file, like the get_article and get_video.py, are to be run in a batch file form, thought it NEED not to be run manually, you could run it in the same folder where there is "exercise.txt". Note: running this depends on: selenium, urllib.request lib and docx, and everyone's chromdriver may vary, so you might need to change: the location of your chromedriver before running. Otherwise there will be errors. Also, I have not yet been able to dismiss the browser in the code (since I am afraid there will be errors), if wished, I may update a final form in this repository.